Move resource definition functions to own file
